// src/types/store.types.ts
export interface StoreType {
  "ACTUAL CLIENT STORE ID": string;
  "BUSINESS / BRAND NAME": string;
  "ADDRESS 1": string;
  "ADDRESS 2": string;
  "MAIN PHONE NO.": string;
  POSTCODE: string;
  STATE: string;
  "STORE EMAIL ID": string;
  CITY: string;
  LOCALITY: string;
  LATITUDE: string;
  LONGITUDE: string;
  LANDMARK: string;
}